ノートブックとページをパート構成にリネームし、app.pyを動作確認用に簡素化
