+++ day2/pages/part1_4_naive_agent.py
