+++ day2/pages/part2_1_indexing.py
